edit the story repo

+++ src/main/java/com/storywatpad/backend/repository/StoryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface StoryRepository extends JpaRepository<Story, Integer> {
+public interface StoryRepository extends JpaRepository<Story, Long> {
     List<Story> findTop5ByIsHiddenFalseOrderByCreatedAtDesc();
 }
